# AuthService/models/group_list.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// GroupList A list of groups
// swagger:model GroupList
type GroupList struct {

	// list
	// Required: true
	// Min Items: 1
	List []string `json:"list"`
}

// Validate validates this group list
func (m *GroupList) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateList(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *GroupList) validateList(formats strfmt.Registry) error {

	if err := validate.Required("list", "body", m.List); err != nil {
		return err
	}

	iListSize := int64(len(m.List))

	if err := validate.MinItems("list", "body", iListSize, 1); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *GroupList) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *GroupList) UnmarshalBinary(b []byte) error {
	var res GroupList
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}